Version 1.5.0 AutoGUI working Heaeder-Only version of GetSet for AutoGUI Install scripts

+++ examples/autogui_header_only/GetSetAutoGui.hxx
@@ -105,7 +105,7 @@ inline void trim(std::string& str, const std::string& t = " \t")
 inline std::string enumNth(int n, const std::string& choices)
 {
 	std::vector<std::string> c=stringToVector<std::string>(choices,';');
-	if (n>=0 && n<c.size())
+	if (n>=0 && n<(int)c.size())
 		return c[n];
 	else
 		return "";
